Multiple changes
Set SO_REUSEADDR on the socket, so other apps can send/listen on the
SSDP channel. Practical if you use for ex. VLC on the machine that
runs dlnaproxy.

Bugfix: the scheduler doesn't want to run if it's not the last thing
called. Hacked a fix...

Dep: removed syslog deps.